Properly define CFLAGS and remove some warnings.
There is a lot of more warnings with -Wall; for now
just disable the flag.

+++ lib/status.c
@@ -1,10 +1,10 @@
 #include "status.h"
 
-unsigned get_progress_percentage(struct progress *prog) {
+unsigned long get_progress_percentage(struct progress *prog) {
 	return ((double) prog->current / prog->total) * 100;
 }
 
-unsigned inc_and_get_progress_percentage(struct progress *prog) {
+unsigned long inc_and_get_progress_percentage(struct progress *prog) {
 	prog->current++;
 	return get_progress_percentage(prog);
 }
